fix: move redesign changes to correct monorepo paths
Apply home page layout redesign to apps/web/pages/index.vue and
apps/web/components/atoms/CountdownDigits.vue instead of root paths.

Changes:
- Update apps/web/pages/index.vue with redesigned layout (hero timer,
  inline presets, dashboard grid, simplified dependencies)
- Update apps/web/components/atoms/CountdownDigits.vue with responsive
  sizing (w-[2em]/md:w-[2.5em], py-3/md:py-4)
- Remove incorrect root files (pages/index.vue, components/atoms/CountdownDigits.vue)
