+++ apps/web/components/atoms/CountdownDigits.vue
@@ -1,9 +1,9 @@
 <template>
-  <div class="flex gap-x-2">
+  <div class="flex gap-x-1 md:gap-x-2">
     <span
       v-for="(digit, index) in splittedDigits"
       :key="index"
-      class="flex-1 bg-base-200 rounded-2xl flex items-center justify-center py-4 text-inherit font-rajdhani font-bold text-base-content shadow-inner"
+      class="w-[2em] md:w-[2.5em] bg-base-200 rounded-2xl flex items-center justify-center py-3 md:py-4 font-rajdhani font-bold text-base-content shadow-inner"
     >
       {{ digit }}
     </span>
